Merubah penulisan kode menjadi file .py terpisah (Splitting code into multiple files)
- Menyerderhanakan penulisan kode
- Memudahkan penambahan source code baru ke dalam proyek
- Kemudahan kolaborasi developer
