+++ main.py
@@ -1,17 +1,21 @@
 # Pustaka Rich
-from rich import print  # Warna Teks
-from rich.progress import track  # Progress Bar
-import time  # Time Progress Bar
-from rich.panel import Panel  # Menyorot Teks
-from rich.console import Console  # Menyorot Teks
+from rich.console import Console # Menyorot Teks
+from rich.panel import Panel # Menyorot Teks
 from rich.prompt import Prompt # Input interaktif
+from rich import print # Warna Teks
+
+# Impor fungsi fiturSatu-fiturTiga
+from features1 import fiturSatu
+from features2 import fiturDua
+from features3 import fiturTiga
+
+# Impor fungsi progressBar
+from utils import progressBar
+
+# Buat objek Console dari pustaka Rich
+console = Console()
 
 # Fungsi
-def progressBar():
-    for _ in track(range(2), description="[bold green]Memuat..."):
-        time.sleep(0.25)
-    print("")
-
 def menuUtama():
     console.print(Panel("Menu Utama", style="bold bright_white", width=14))
     console.print("[bold bright_white]1. Yo-Managements[/bold bright_white]")
@@ -21,43 +25,11 @@ def menuUtama():
     pilihan = Prompt.ask("[bold bright_green]Pilih menu yang ingin Anda akses (1-3), atau pilih 4 untuk keluar[/bold bright_green]", choices=["1", "2", "3", "4"])
     return pilihan
 
-def fiturSatu():
-    console.print(Panel("Yo-Managements", style="bold bright_cyan", width=18))
-    kembali = Prompt.ask("[bold bright_blue]Ketik 'kembali' untuk kembali ke menu utama[/bold bright_blue]", choices=["kembali"])
-    if kembali.lower() == "kembali":
-        progressBar()
-        return True
-    return False
-
-def fiturDua():
-    console.print(Panel("Perhitungan Tabungan", style="bold bright_cyan", width=24))
-    kembali = Prompt.ask("[bold bright_blue]Ketik 'kembali' untuk kembali ke menu utama[/bold bright_blue]", choices=["kembali"])
-    if kembali.lower() == "kembali":
-        progressBar()
-        return True
-    return False
-
-def fiturTiga():
-    console.print(Panel("Bantuan Pengguna", style="bold bright_cyan", width=20))
-    kembali = Prompt.ask("[bold bright_blue]Ketik 'kembali' untuk kembali ke menu utama[/bold bright_blue]", choices=["kembali"])
-    if kembali.lower() == "kembali":
-        progressBar()
-        return True
-    return False
-
-# Inisialisasi objek Console dari pustaka Rich
-console = Console()
-
-""" 
-Halaman Utama | Ayona 
-Versi 0.1
-"""""
-greeting = "Selamat datang di aplikasi Ayona!"
-console.print(Panel(f"\n\t\t{greeting}\t\t\n", title="Sistem Ayona", title_align="right", style="bold bright_blue", width=64))
-progressBar()
-
-# Program utama
 def main():
+    greeting = "Selamat datang di aplikasi Ayona!"
+    console.print(Panel(f"\n\t\t{greeting}\t\t\n", title="Sistem Ayona", title_align="right", style="bold bright_blue", width=64))
+    progressBar()
+
     while True:
         pilihan = menuUtama()
 
@@ -82,7 +54,6 @@ def main():
             return False
 
 """
-if __name__ == "__main__":
 Memeriksa apakah skrip sedang dijalankan langsung (bukan diimpor).
 Jika benar, Python akan memanggil fungsi main().
 Jika salah, Python tidak akan memanggil fungsi main().
